+++ lib/client.js
@@ -1,48 +1,475 @@
 "use strict";
-var MatrixHttpApi = require("./http-api");
+var httpApi = require("./http-api");
 var MatrixEvent = require("./models/event").MatrixEvent;
+var utils = require("./utils");
 
 // TODO:
 // Internal: rate limiting
 
 /*
  * Construct a Matrix Client.
- * @param {Object} credentials The credentials for this client
- * @param {Object} config The config (if any) for this client.
- *  Valid config params include:
- *      noUserAgent: true // to avoid warnings whilst setting UA headers
- *      debug: true // to use console.err() style debugging from the lib
- * @param {Object} store The data store (if any) for this client.
- * @param {Function} request The request fn to use.
+ * @param {Object} opts The configuration options for this client.
+ * @param {string} opts.baseUrl Required. The base URL to the client-server HTTP API.
+ * @param {Function} opts.request Required. The function to invoke for HTTP requests.
+ * @param {boolean} opts.usePromises True to use promises rather than callbacks.
+ * @param {string} opts.accessToken The access_token for this user.
+ * @param {string} opts.userId The user ID for this user.
+ * @param {Object} opts.store The data store to use. See {@link store/memory}.
  */
-function MatrixClient(credentials, config, store, request) {
-    if (typeof credentials === "string") {
-        credentials = {
-            "baseUrl": credentials
-        };
-    }
-    var requiredKeys = [
-        "baseUrl"
-    ];
-    for (var i = 0; i < requiredKeys.length; i++) {
-        if (!credentials.hasOwnProperty(requiredKeys[i])) {
-            throw new Error("Missing required key: " + requiredKeys[i]);
-        }
-    }
-    this.config = config;
-    this.credentials = credentials;
-    this.store = store;
-
+function MatrixClient(opts) {
+    utils.checkObjectHasKeys(opts, ["baseUrl", "request"]);
+    utils.checkObjectHasNoAdditionalKeys(opts,
+        ["baseUrl", "request", "usePromises", "accessToken", "userId", "store"]
+    );
+    
+    this.store = opts.store || null;
     // track our position in the overall eventstream
     this.fromToken = undefined;
     this.clientRunning = false;
-    this._http = new MatrixHttpApi(credentials, config, request);
+
+    var httpOpts = {
+        baseUrl: opts.baseUrl,
+        accessToken: opts.accessToken,
+        request: opts.request,
+        prefix: httpApi.PREFIX_V1
+    };
+    this.credentials = {
+        userId: (opts.userId || null)
+    };
+    this._http = new httpApi.MatrixHttpApi(httpOpts);
 }
 MatrixClient.prototype = {
-    isLoggedIn: function() {
-        return this.credentials.accessToken !== undefined &&
-            this.credentials.userId !== undefined;
+
+    // Room operations
+    // ===============
+
+    createRoom: function(options, callback) {
+        // valid options include: room_alias_name, visibility, invite
+        return this._http.authedRequest(
+            callback, "POST", "/createRoom", undefined, options
+        );
     },
+
+    joinRoom: function(roomIdOrAlias, callback) {
+        var path = utils.encodeUri("/join/$roomid", { $roomid: roomIdOrAlias});
+        return this._http.authedRequest(callback, "POST", path, undefined, {});
+    },
+
+    setRoomName: function(roomId, name, callback) {
+        return this.sendStateEvent(roomId, "m.room.name", {name: name},
+                                   undefined, callback);
+    },
+
+    setRoomTopic: function(roomId, topic, callback) {
+        return this.sendStateEvent(roomId, "m.room.topic", {topic: topic},
+                                   undefined, callback);
+    },
+
+    setPowerLevel: function(roomId, userId, powerLevel, event, callback) {
+        var content = {
+            users: {}
+        };
+        if (event && event.type === "m.room.power_levels") {
+            content = event.content;
+        }
+        content.users[userId] = powerLevel;
+        var path = utils.encodeUri("/rooms/$roomId/state/m.room.power_levels", {
+            $roomId: roomId
+        });
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, content
+        );
+    },
+
+    getStateEvent: function(roomId, eventType, stateKey, callback) {
+        var pathParams = {
+            $roomId: roomId,
+            $eventType: eventType,
+            $stateKey: stateKey
+        };
+        var path = utils.encodeUri("/rooms/$roomId/state/$eventType", pathParams);
+        if (stateKey !== undefined) {
+            path = utils.encodeUri(path + "/$stateKey", pathParams);
+        }
+        return this._http.authedRequest(
+            callback, "GET", path
+        );
+    },
+
+    sendStateEvent: function(roomId, eventType, content, stateKey, 
+                             callback) {
+        var pathParams = {
+            $roomId: roomId,
+            $eventType: eventType,
+            $stateKey: stateKey
+        };
+        var path = utils.encodeUri("/rooms/$roomId/state/$eventType", pathParams);
+        if (stateKey !== undefined) {
+            path = utils.encodeUri(path + "/$stateKey", pathParams);
+        }
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, content
+        );
+    },
+
+    sendEvent: function(roomId, eventType, content, txnId, callback) {
+        if (utils.isFunction(txnId)) { callback = txnId; txnId = undefined; }
+
+        if (!txnId) {
+            txnId = "m" + new Date().getTime();
+        }
+
+        var path = utils.encodeUri("/rooms/$roomId/send/$eventType/$txnId", {
+            $roomId: roomId,
+            $eventType: eventType,
+            $txnId: txnId
+        });
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, content
+        );
+    },
+
+    sendMessage: function(roomId, content, txnId, callback) {
+        if (utils.isFunction(txnId)) { callback = txnId; txnId = undefined; }
+        return this.sendEvent(
+            roomId, "m.room.message", content, txnId, callback
+        );
+    },
+
+    sendTextMessage: function(roomId, body, txnId, callback) {
+        var content = {
+             msgtype: "m.text",
+             body: body
+        };
+        return this.sendMessage(roomId, content, txnId, callback);
+    },
+
+    sendEmoteMessage: function(roomId, body, txnId, callback) {
+        var content = {
+             msgtype: "m.emote",
+             body: body
+        };
+        return this.sendMessage(roomId, content, txnId, callback);
+    },
+
+    sendImageMessage: function(roomId, url, info, text, callback) {
+        if (utils.isFunction(text)) { callback = text; text = undefined; }
+        if (!text) { text = "Image"; }
+        var content = {
+             msgtype: "m.image",
+             url: url,
+             info: info,
+             body: text
+        };
+        return this.sendMessage(roomId, content, callback);
+    },
+
+    sendHtmlMessage: function(roomId, body, htmlBody, callback) {
+        var content = {
+            msgtype: "m.text",
+            format: "org.matrix.custom.html",
+            body: body,
+            formatted_body: htmlBody
+        };
+        return this.sendMessage(roomId, content, callback);
+    },
+
+    sendTyping: function(roomId, isTyping, timeoutMs, callback) {
+        var path = utils.encodeUri("/rooms/$roomId/typing/$userId", {
+            $roomId: roomId,
+            $userId: this.credentials.userId
+        });
+        var data = {
+            typing: isTyping
+        };
+        if (isTyping) {
+            data.timeout = timeoutMs ? timeoutMs : 20000;
+        }
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, data
+        );
+    },
+
+    redactEvent: function(roomId, eventId, callback) {
+        var path = utils.encodeUri("/rooms/$roomId/redact/$eventId", {
+            $roomId: roomId,
+            $eventId: eventId
+        });
+        return this._http.authedRequest(callback, "POST", path, undefined, {});
+    },
+
+    invite: function(roomId, userId, callback) {
+        return this._membershipChange(roomId, userId, "invite", undefined,
+            callback);
+    },
+
+    leave: function(roomId, callback) {
+        return this._membershipChange(roomId, undefined, "leave", undefined,
+            callback);
+    },
+
+    ban: function(roomId, userId, reason, callback) {
+        return this._membershipChange(roomId, userId, "ban", reason,
+            callback);
+    },
+
+    unban: function(roomId, userId, callback) {
+        // unbanning = set their state to leave
+        return this._setMembershipState(
+            roomId, userId, "leave", undefined, callback
+        );
+    },
+
+    kick: function(roomId, userId, reason, callback) {
+        return this._setMembershipState(
+            roomId, userId, "leave", reason, callback
+        );
+    },
+
+    _setMembershipState: function(roomId, userId, membershipValue, reason, 
+                            callback) {
+        if (utils.isFunction(reason)) { callback = reason; reason = undefined; }
+
+        var path = utils.encodeUri(
+            "/rooms/$roomId/state/m.room.member/$userId",
+            { $roomId: roomId, $userId: userId}
+        );
+
+        return this._http.authedRequest(callback, "PUT", path, undefined, {
+            membership: membershipValue,
+            reason: reason
+        });
+    },
+
+    _membershipChange: function(roomId, userId, membership, reason, 
+                                callback) {
+        if (utils.isFunction(reason)) { callback = reason; reason = undefined; }
+
+        var path = utils.encodeUri("/rooms/$room_id/$membership", {
+            $room_id: roomId,
+            $membership: membership
+        });
+        return this._http.authedRequest(
+            callback, "POST", path, undefined, {
+                user_id: userId,  // may be undefined e.g. on leave
+                reason: reason
+            }
+        );
+    },
+
+    // Profile operations
+    // ==================
+
+    getProfileInfo: function(userId, info, callback) {
+        if (utils.isFunction(info)) { callback = info; info = undefined; }
+
+        var path = info ?
+        utils.encodeUri("/profile/$userId/$info",
+                 { $userId: userId, $info: info }) :
+        utils.encodeUri("/profile/$userId",
+                 { $userId: userId });
+        return this._http.authedRequest(callback, "GET", path);
+    },
+
+    setProfileInfo: function(info, data, callback) {
+        var path = utils.encodeUri("/profile/$userId/$info", {
+            $userId: this.credentials.userId,
+            $info: info
+        });
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, data
+        );
+    },
+
+    setDisplayName: function(name, callback) {
+        return this.setProfileInfo(
+            "displayname", { displayname: name }, callback
+        );
+    },
+
+    setAvatarUrl: function(url, callback) {
+        return this.setProfileInfo(
+            "avatar_url", { avatar_url: url }, callback
+        );
+    },
+
+    getThreePids: function(creds, bind, callback) {
+        var path = "/account/3pid";
+        return this._http.authedRequestWithPrefix(
+            callback, "GET", path, undefined, undefined, httpApi.PREFIX_V2_ALPHA
+        );
+    },
+
+    addThreePid: function(creds, bind, callback) {
+        var path = "/account/3pid";
+        var data = {
+            'threePidCreds': creds,
+            'bind': bind
+        };
+        return this._http.authedRequestWithPrefix(
+            callback, "POST", path, qps, data, httpApi.PREFIX_V2_ALPHA
+        );
+    },
+
+    setPresence: function(presence, callback) {
+        var path = utils.encodeUri("/presence/$userId/status", {
+            $userId: this.credentials.userId
+        });
+        var validStates = ["offline", "online", "unavailable"];
+        if (validStates.indexOf(presence) == -1) {
+            throw new Error("Bad presence value: " + presence);
+        }
+        var content = {
+            presence: presence
+        };
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, content
+        );
+    },
+
+    // Public (non-authed) operations
+    // ==============================
+
+    publicRooms: function(callback) {
+        return this._http.request(callback, "GET", "/publicRooms");
+    },
+
+    registerFlows: function(callback) {
+        return this._http.request(callback, "GET", "/register");
+    },
+
+    loginFlows: function(callback) {
+        return this._http.request(callback, "GET", "/login");
+    },
+
+    resolveRoomAlias: function(roomAlias, callback) {
+        var path = utils.encodeUri("/directory/room/$alias", {$alias: roomAlias});
+        return this._http.request(callback, "GET", path);
+    },
+
+    // Syncing operations
+    // ==================
+
+    initialSync: function(limit, callback) {
+        var params = {
+            limit: limit
+        };
+        return this._http.authedRequest(
+            callback, "GET", "/initialSync", params
+        );
+    },
+
+    roomInitialSync: function(roomId, limit, callback) {
+        if (utils.isFunction(limit)) { callback = limit; limit = undefined; }
+        var path = utils.encodeUri("/rooms/$roomId/initialSync",
+            {$roomId: roomId}
+        );
+        if (!limit) {
+            limit = 30;
+        }
+        return this._http.authedRequest(
+            callback, "GET", path, { limit: limit }
+        );
+    },
+
+    roomState: function(roomId, callback) {
+        var path = utils.encodeUri("/rooms/$roomId/state", {$roomId: roomId});
+        return this._http.authedRequest(callback, "GET", path);
+    },
+
+    scrollback: function(roomId, from, limit, callback) {
+        if (utils.isFunction(limit)) { callback = limit; limit = undefined; }
+        var path = utils.encodeUri("/rooms/$roomId/messages", {$roomId: roomId});
+        if (!limit) {
+            limit = 30;
+        }
+        var params = {
+            from: from,
+            limit: limit,
+            dir: 'b'
+        };
+        return this._http.authedRequest(callback, "GET", path, params);
+    },
+
+    eventStream: function(from, timeout, callback) {
+        if (utils.isFunction(timeout)) { callback = timeout; timeout = undefined;}
+        if (!timeout) {
+            timeout = 30000;
+        }
+
+        var params = {
+            from: from,
+            timeout: timeout
+        };
+        return this._http.authedRequest(callback, "GET", "/events", params);
+    },
+
+    // Registration/Login operations
+    // =============================
+
+    login: function(loginType, data, callback) {
+        data.type = loginType;
+        return this._http.authedRequest(
+            callback, "POST", "/login", undefined, data
+        );
+    },
+
+    register: function(loginType, data, callback) {
+        data.type = loginType;
+        return this._http.authedRequest(
+            callback, "POST", "/register", undefined, data
+        );
+    },
+
+    loginWithPassword: function(user, password, callback) {
+        return this.login("m.login.password", {
+            user: user,
+            password: password
+        }, callback);
+    },
+
+    // Push operations
+    // ===============
+
+    pushRules: function(callback) {
+        return this._http.authedRequest(callback, "GET", "/pushrules/");
+    },
+
+    addPushRule: function(scope, kind, ruleId, body, callback) {
+        // NB. Scope not uri encoded because devices need the '/'
+        var path = utils.encodeUri("/pushrules/" + scope + "/$kind/$ruleId", {
+            $kind: kind,
+            $ruleId: ruleId
+        });
+        return this._http.authedRequest(
+            callback, "PUT", path, undefined, body
+        );
+    },
+
+    deletePushRule: function(scope, kind, ruleId, callback) {
+        // NB. Scope not uri encoded because devices need the '/'
+        var path = utils.encodeUri("/pushrules/" + scope + "/$kind/$ruleId", {
+            $kind: kind,
+            $ruleId: ruleId
+        });
+        return this._http.authedRequest(callback, "DELETE", path);
+    },
+
+    // VoIP operations
+    // ===============
+
+    turnServer: function(callback) {
+        return this._http.authedRequest(callback, "GET", "/voip/turnServer");
+    },
+
+    isLoggedIn: function() {
+        return this._http.opts.accessToken !== undefined
+    },
+
+
+
+
+
 
     // Higher level APIs
     // =================
@@ -190,14 +617,14 @@ MatrixClient.prototype = {
                     };
                     // intercept the results and put them into our store
                     self.store.setPresenceEvents(
-                        map(data.presence, eventMapper)
+                        utils.map(data.presence, eventMapper)
                     );
                     for (i = 0; i < data.rooms.length; i++) {
                         self.store.setStateEvents(
-                            map(data.rooms[i].state, eventMapper)
+                            utils.map(data.rooms[i].state, eventMapper)
                         );
                         self.store.setEvents(
-                            map(data.rooms[i].messages.chunk, eventMapper)
+                            utils.map(data.rooms[i].messages.chunk, eventMapper)
                         );
                     }
                 }
@@ -252,7 +679,7 @@ MatrixClient.prototype = {
             }
 
             if (self.store) {
-                self.store.setEvents(map(data.chunk,
+                self.store.setEvents(utils.map(data.chunk,
                     function(event) {
                         return new MatrixEvent(event);
                     }
@@ -277,14 +704,6 @@ MatrixClient.prototype = {
     stopClient: function() {
         this.clientRunning = false;
     },
-};
-
-var map = function(array, fn) {
-    var results = new Array(array.length);
-    for (var i = 0; i < array.length; i++) {
-        results[i] = fn(array[i]);
-    }
-    return results;
 };
 
 /**
