+++ lib/matrix.js
@@ -20,14 +20,23 @@ module.exports.request = function(r) {
     request = r;
 };
 
-/**
- * Create a new Matrix Client.
- * @param {Object} credentials The Matrix credentials to use.
- * @param {Object} config The config options for the client
- * @param {Store} store The type of store to use.
- * @return {MatrixClient} A new Matrix Client
+/*
+ * Construct a Matrix Client. Identical to {@link client/MatrixClient} except
+ * the 'request' option is already specified.
+ * @param {(Object|string)} opts The configuration options for this client. If
+ * this is a string, it is assumed to be the base URL.
+ * @param {string} opts.baseUrl The base URL to the client-server HTTP API.
+ * @param {boolean} opts.usePromises True to use promises rather than callbacks.
+ * @param {string} opts.accessToken The access_token for this user.
+ * @param {string} opts.userId The user ID for this user.
  */
-module.exports.createClient = function(credentials, config, store) {
-    return new module.exports.MatrixClient(credentials, config, store, request);
+module.exports.createClient = function(opts) {
+    if (typeof opts === "string") {
+        opts = {
+            "baseUrl": opts
+        };
+    }
+    opts.request = request;
+    return new module.exports.MatrixClient(opts);
 };
